requirements.txt
PyGithub==2.1.1
Flask==3.0.0
pandas==2.1.4
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
